+++ src/components/charts/common/AdmissionCountsByType.js
@@ -33,7 +33,7 @@ import {
   COLORS,
   COLORS_FIVE_VALUES,
 } from "../../../assets/scripts/constants/colors";
-import { configureDownloadButtons } from "../../../assets/scripts/utils/downloads";
+import { configureDownloadButtons } from "../../../utils/downloads/downloads";
 import {
   filterDatasetByDistrict,
   filterDatasetBySupervisionType,
