+++ src/components/charts/common/bars/utils.js
@@ -24,7 +24,7 @@ import toInteger from "lodash/fp/toInteger";
 import values from "lodash/fp/values";
 import upperCase from "lodash/fp/upperCase";
 
-import { configureDownloadButtons } from "../../../../assets/scripts/utils/downloads";
+import { configureDownloadButtons } from "../../../../utils/downloads/downloads";
 
 /**
  * Casts arguments to integer and sum it.
