+++ src/components/charts/facilities/ReincarcerationCountOverTime.js
@@ -27,7 +27,7 @@ import toInteger from "lodash/fp/toInteger";
 import values from "lodash/fp/values";
 
 import { COLORS } from "../../../assets/scripts/constants/colors";
-import { configureDownloadButtons } from "../../../assets/scripts/utils/downloads";
+import { configureDownloadButtons } from "../../../utils/downloads/downloads";
 import { filterDatasetByDistrict } from "../../../utils/charts/dataFilters";
 import {
   getGoalForChart,
