+++ src/components/charts/programming/FtrReferralCountByMonth.js
@@ -26,7 +26,7 @@ import toInteger from "lodash/fp/toInteger";
 import { groupByMonth } from "../common/bars/utils";
 
 import { COLORS } from "../../../assets/scripts/constants/colors";
-import { configureDownloadButtons } from "../../../assets/scripts/utils/downloads";
+import { configureDownloadButtons } from "../../../utils/downloads/downloads";
 import {
   filterDatasetBySupervisionType,
   filterDatasetByDistrict,
