+++ src/utils/downloads/getFilters.js
@@ -1,8 +1,8 @@
 import pipe from "lodash/fp/pipe";
 import find from "lodash/fp/find";
 import get from "lodash/fp/get";
-import { humanReadableTitleCase } from "../../../utils/transforms/labels";
-import { SUPERVISION_LEVELS } from "../../../views/tenants/constants/filterOptions";
+import { humanReadableTitleCase } from "../transforms/labels";
+import { SUPERVISION_LEVELS } from "../../views/tenants/constants/filterOptions";
 
 function formatMetricPeriodMonthsFilter(metricPeriodMonths) {
   switch (metricPeriodMonths) {
