+++ src/utils/downloads/getTimeStamp.js
@@ -14,6 +14,10 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
+import moment from "moment";
 
-import "../styles/index.scss";
-import "./charts";
+function getTimeStamp() {
+  return moment().format("MM-DD-YYYY-hh-mm-ss-A");
+}
+
+export default getTimeStamp;
